[expo-router] fix webModalStyle TypeScript error in Stack.Screen options
`ExtendedStackNavigationOptions` (which defines `webModalStyle`) was only
defined in `StackClient.tsx`, so it was never applied to `Stack.Screen`'s
`options` prop type. Users trying to pass `webModalStyle` in a layout file
got a TypeScript error: "Object literal may only specify known properties".

Move the type to `stack-utils/StackScreen.tsx` where `StackBaseScreenProps`
is constructed, so `options` now correctly accepts `webModalStyle`. Re-export
from `StackClient.tsx` so existing imports in the web modal files continue
to work unchanged.

Fixes #44485
