+++ packages/expo-router/CHANGELOG.md
@@ -19,6 +19,7 @@
 
 ### 🐛 Bug fixes
 
+- Fix TypeScript error when passing `webModalStyle` to `Stack.Screen` options by moving `ExtendedStackNavigationOptions` to `stack-utils` so it applies to the `options` prop type. ([#44508](https://github.com/expo/expo/pull/44508) by [@Nedunchezhiyan-M](https://github.com/Nedunchezhiyan-M))
 - Fix pinned `react-navigation` dependencies ([#43456](https://github.com/expo/expo/pull/43456) by [@kitten](https://github.com/kitten))
 - Fix zoom transition to prefetched routes ([#43852](https://github.com/expo/expo/pull/43852) by [@Ubax](https://github.com/Ubax))
 - Fix `Stack.Protected` guard not applying to index routes. ([#43769](https://github.com/expo/expo/pull/43769) by [@alanjhughes](https://github.com/alanjhughes))
