+++ packages/expo-router/src/layouts/StackClient.tsx
@@ -15,6 +15,7 @@ import {
 } from '../navigationParams';
 import { SingularOptions, getSingularId } from '../useScreens';
 import {
+  type ExtendedStackNavigationOptions,
   type StackScreenProps,
   StackHeader,
   StackScreen,
@@ -37,10 +38,7 @@ import {
   StackNavigationState,
   type RouteProp,
 } from '../react-navigation/native';
-import {
-  NativeStackNavigationEventMap,
-  NativeStackNavigationOptions,
-} from '../react-navigation/native-stack';
+import { NativeStackNavigationEventMap } from '../react-navigation/native-stack';
 import { isChildOfType } from '../utils/children';
 import { Protected } from '../views/Protected';
 
@@ -48,49 +46,9 @@ type GetId = NonNullable<RouterConfigOptions['routeGetIdList'][string]>;
 
 const NativeStackNavigator = createNativeStackNavigator().Navigator;
 
-/**
- * We extend NativeStackNavigationOptions with our custom props
- * to allow for several extra props to be used on web, like modalWidth
- */
-export type ExtendedStackNavigationOptions = NativeStackNavigationOptions & {
-  webModalStyle?: {
-    /**
-     * Override the width of the modal (px or percentage). Only applies on web platform.
-     * @platform web
-     */
-    width?: number | string;
-    /**
-     * Override the height of the modal (px or percentage). Applies on web desktop.
-     * @platform web
-     */
-    height?: number | string;
-    /**
-     * Minimum height of the desktop modal (px or percentage). Overrides the default 640px clamp.
-     * @platform web
-     */
-    minHeight?: number | string;
-    /**
-     * Minimum width of the desktop modal (px or percentage). Overrides the default 580px.
-     * @platform web
-     */
-    minWidth?: number | string;
-    /**
-     * Override the border of the desktop modal (any valid CSS border value, e.g. '1px solid #ccc' or 'none').
-     * @platform web
-     */
-    border?: string;
-    /**
-     * Override the overlay background color (any valid CSS color or rgba/hsla value).
-     * @platform web
-     */
-    overlayBackground?: string;
-    /**
-     * Override the modal shadow filter (any valid CSS filter value, e.g. 'drop-shadow(0 4px 8px rgba(0,0,0,0.1))' or 'none').
-     * @platform web
-     */
-    shadow?: string;
-  };
-};
+// Re-exported for the web modal files that import ExtendedStackNavigationOptions from this module.
+// The type is defined in stack-utils/StackScreen so it's available to Stack.Screen options.
+export type { ExtendedStackNavigationOptions };
 
 const RNStack = withLayoutContext<
   ExtendedStackNavigationOptions,
