+++ packages/expo-router/src/layouts/stack-utils/index.tsx
@@ -11,6 +11,7 @@ export {
   StackScreen,
   appendScreenStackPropsToOptions,
   validateStackPresentation,
+  type ExtendedStackNavigationOptions,
   type StackScreenProps,
 } from './StackScreen';
